Genome v6.1
Lots of changes to this one. I realized a few things. First, the
modified neural net actually doesn't talk to coved's. I forgot to remake
the network after modifying it. Luckily this was a relativly easy fix
after realizing what was wrong. I started work on the crossover function
and finished most of it. I will start testing the crossover tomorrow. I
also started work on the fitness function and believe I have a good way
to select genome based on their fitness. I hope by the end of the night
to have this completed by the end of the night.

+++ src/com/riskybusiness/genetic/NeuronGene.java
@@ -35,7 +35,7 @@ public class NeuronGene implements Serializable
     //Represents whether the neuron is recurrent
     private boolean neuronRecurrent;
     //Represents the weight of the neuron
-    private float   neuronActivationResponse;
+    private double   neuronActivationResponse;
     //Represents the layer of the neuron(used in the genome class)
     //private String  neuronLayer;
     //private int     innovationNum;
@@ -48,12 +48,12 @@ public class NeuronGene implements Serializable
         neuronID = -1;
         neuronType = "SIGMOID";
         neuronRecurrent = false;
-        neuronActivationResponse = 0;
+        neuronActivationResponse = 0.0;
         neuronLayer = 1;
     }
 
     //Create a neuron gene
-    public NeuronGene(int id, String type, boolean recur, float activate, int layer)
+    public NeuronGene(int id, String type, boolean recur, double activate, int layer)
     {
         neuronID = id;
         neuronType = type;
@@ -93,13 +93,13 @@ public class NeuronGene implements Serializable
     }
     
     //Get the activation response of the neuron
-    public float getActivationResponse()
+    public double getActivationResponse()
     {
         return this.neuronActivationResponse;
     }
     
     //Set the activation response of the neuron
-    public void setActivationResponse(float activate)
+    public void setActivationResponse(double activate)
     {
         this.neuronActivationResponse = activate;
     }
